refact on collection factories and create block factory
